backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
sentence-transformers
numpy
python-multipart